Add files via upload
Student management system:
This project deals with the basic tasks involved for student management system like creating a new record for student, displaying existing records, modifying an existing record, search for an existing record, delete an existing record. Here file handling is used so a file containing the details of all the students will be  created and all operations like search, modify, delete, add new, delete  on the records will be done on this file and thus this file will have  the values dynamically updated depending on the operations performed on it.